@@ -24,11 +24,12 @@ from launch.substitutions import LaunchConfiguration
 import driving_log_replayer.launch_common
 
 
-def generate_launch_description():
+def generate_launch_description() -> launch.LaunchDescription:
     launch_arguments = driving_log_replayer.launch_common.get_driving_log_replayer_common_argument()
     launch_arguments.append(DeclareLaunchArgument("sensing", default_value="false"))
     autoware_launch = driving_log_replayer.launch_common.get_autoware_launch(
-        sensing=LaunchConfiguration("sensing"), localization="false"
+        sensing=LaunchConfiguration("sensing"),
+        localization="false",
     )
     rviz_node = driving_log_replayer.launch_common.get_rviz("perception.rviz")
     evaluator_node = driving_log_replayer.launch_common.get_evaluator_node("perception")
@@ -48,7 +49,7 @@ def generate_launch_description():
         [
             "--remap",
             "/sensing/lidar/concatenated/pointcloud:=/driving_log_replayer/concatenated/pointcloud",
-        ]
+        ],
     )
 
     player_normal = ExecuteProcess(
@@ -79,5 +80,5 @@ def generate_launch_description():
             player_normal,
             player_remap,
             recorder,
-        ]
+        ],
     )
