@@ -17,11 +17,12 @@ import launch
 import driving_log_replayer.launch_common
 
 
-def generate_launch_description():
+def generate_launch_description() -> launch.LaunchDescription:
     launch_arguments = driving_log_replayer.launch_common.get_driving_log_replayer_common_argument()
     fitter_launch = driving_log_replayer.launch_common.get_map_height_fitter(launch_service="true")
     autoware_launch = driving_log_replayer.launch_common.get_autoware_launch(
-        perception="false", pose_source="yabloc"
+        perception="false",
+        pose_source="yabloc",
     )
     rviz_node = driving_log_replayer.launch_common.get_rviz("localization.rviz")
     evaluator_node = driving_log_replayer.launch_common.get_evaluator_node("yabloc")
@@ -48,5 +49,5 @@ def generate_launch_description():
             evaluator_node,
             recorder,
             player,
-        ]
+        ],
     )
