@@ -23,7 +23,7 @@ import simplejson as json
 
 
 class PerceptionLoadDatabaseResult:
-    def __init__(self, result_root_directory, scenario_path):
+    def __init__(self, result_root_directory: str, scenario_path: str) -> None:
         analyzer = PerceptionAnalyzer3D.from_scenario(
             result_root_directory,
             scenario_path,
@@ -44,7 +44,7 @@ class PerceptionLoadDatabaseResult:
             json.dump(database_metrics, f)
 
 
-def main():
+def main() -> None:
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "-r",
@@ -60,7 +60,8 @@ def main():
     )
     args = parser.parse_args()
     PerceptionLoadDatabaseResult(
-        os.path.expandvars(args.result_root_directory), os.path.expandvars(args.scenario_path)
+        os.path.expandvars(args.result_root_directory),
+        os.path.expandvars(args.scenario_path),
     )
 
 
