@@ -13,7 +13,7 @@ except Exception:  # noqa
 CONTEXT_SETTINGS = {"help_option_names": ["-h", "--help"]}
 
 
-def main():
+def main() -> None:
     cmd.add_command(configuration)
     cmd.add_command(simulation)
     cmd()
@@ -28,9 +28,14 @@ def print_version(ctx, param, value):  # noqa
 
 @click.group(context_settings=CONTEXT_SETTINGS)
 @click.option(
-    "--version", "-v", is_flag=True, callback=print_version, expose_value=False, is_eager=True
+    "--version",
+    "-v",
+    is_flag=True,
+    callback=print_version,
+    expose_value=False,
+    is_eager=True,
 )
-def cmd():
+def cmd() -> None:
     """
     Command line tool to use driving_log_replayer.
 
