@@ -1,3 +1,3 @@
 class UserError(Exception):
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args, **kwargs) -> None:  # noqa
         super().__init__(*args, **kwargs)
