@@ -9,11 +9,11 @@ import termcolor
 
 
 class DrivingLogReplayerResultViewer:
-    def __init__(self, result_path):
+    def __init__(self, result_path: str) -> None:
         self.__result_path = result_path
 
-    def output(self):
-        print("--------------------------------------------------")
+    def output(self) -> None:
+        print("--------------------------------------------------")  # noqa
         self.__result_json_dict = {}
         with open(self.__result_path) as jsonl_file:
             last_line = jsonl_file.readlines()[-1]
@@ -32,13 +32,12 @@ class DrivingLogReplayerResultViewer:
 
 
 class DrivingLogReplayerResultConverter:
-    def __init__(self, result_path):
+    def __init__(self, result_path: str) -> None:
         self.__result_path = result_path
 
-    def convert(self):
+    def convert(self) -> None:
         output_file_path = Path(self.__result_path).parent.joinpath("result.json")
         if output_file_path.exists():
-            # termcolor.cprint("A json file already exists. Skip convert", "yellow")
             return
         with open(self.__result_path) as jsonl_file:
             result_dict = [json.loads(line) for line in jsonl_file]
@@ -46,7 +45,7 @@ class DrivingLogReplayerResultConverter:
                 json.dump(result_dict, out_file)
 
 
-def display(output_directory: str):
+def display(output_directory: str) -> None:
     regex = os.path.join(os.path.expandvars(output_directory), "**", "result.jsonl")
     result_paths = glob.glob(regex, recursive=True)
     number = 1
@@ -54,13 +53,13 @@ def display(output_directory: str):
     for result_path in natsorted(result_paths):
         result_path_obj = Path(result_path)
         log_directory = result_path_obj.parent
-        print(f"test case {number} / {total} : use case: {log_directory.name}")
+        print(f"test case {number} / {total} : use case: {log_directory.name}")  # noqa
         viewer = DrivingLogReplayerResultViewer(result_path_obj.as_posix())
         viewer.output()
         number = number + 1
 
 
-def convert(output_directory: str):
+def convert(output_directory: str) -> None:
     regex = os.path.join(os.path.expandvars(output_directory), "**", "result.jsonl")
     result_paths = glob.glob(regex, recursive=True)
     for result_path in natsorted(result_paths):
